requirements.txt
APScheduler==3.11.2
Faker==40.13.0
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzlocal==5.3.1
watchdog==6.0.0
